Change jailer's exec_into_new_pid_ns combo of 'unshare' + 'exec' to 'clone'
The exec_into_new_pid_ns() function of the jailer should dispatch a new
process inside a new PID namespace.
It was done using unshare to a new PID namespace, then calling fork to
dispatch a process inside that namespace.
According to the TODO message, it is desirable to replace this combo
with a single call to clone with the CLONE_NEWPID flag.

Besides the actual change inside the function, changed the jailer.md
documentation and the errors in main in order to fit the change.